requirements.txt
annotated-types==0.7.0
anyio==4.9.0
Authlib==1.6.0
blinker==1.9.0
cachelib==0.13.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.4
distro==1.9.0
Flask==3.1.1
Flask-Session==0.8.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
lxml==5.4.0
MarkupSafe==3.0.2
msgspec==0.19.0
nltk==3.9.1
openai==1.86.0
packaging==25.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyPDF2==3.0.1
python-docx==1.1.2
python-dotenv==1.1.0
regex==2024.11.6
requests==2.32.4
sniffio==1.3.1
textblob==0.19.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
Werkzeug==3.1.3
